requirements.txt
flask==2.0.1
werkzeug==2.0.3
ultralytics==8.0.0
opencv-python-headless==4.8.0.74
pillow==9.5.0
numpy==1.24.3